pyproject.toml
[project]
name = "testing"
version = "0.1.0"
description = ""
authors = [
    {name = "Wkdk00",email = "astaningosha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai-whisper (>=20240930,<20240931)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pyannote-audio (>=3.3.2,<4.0.0)",
    "torchaudio (>=2.7.1,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
]

[tool.poetry]
packages = [{include = "testing", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
